+++ Source/OnlineSubsystemAccelByte/Private/OnlineSessionSettingsAccelByte.cpp
@@ -5,6 +5,7 @@
 #include "OnlineSessionSettingsAccelByte.h"
 
 #include "OnlineSubsystemAccelByte.h"
+#include "Math/NumericLimits.h"
 
 constexpr auto DATA_OFFSET = sizeof(uint8);
 
@@ -16,7 +17,7 @@ static void ConvertArrayToBytes(const TArray<FString>& InArray, TArray<uint8>& O
 	OutArray.Add(StaticCast<uint8>(ESessionSettingsAccelByteArrayFieldType::STRINGS));
 
 	TArray<uint8>::SizeType ByteIndex = DATA_OFFSET;
-	for(const auto& String : InArray)
+	for (const auto& String : InArray)
 	{
 		const auto Bytes = String.Len() * sizeof(TCHAR);
 
@@ -36,10 +37,10 @@ static void ConvertArrayToBytes(const TArray<double>& InArray, TArray<uint8>& Ou
 	// Prefix the output bytes with the data type
 	OutArray.Add(StaticCast<uint8>(ESessionSettingsAccelByteArrayFieldType::DOUBLES));
 	// Add enough fields to serialize all of the values in InArray
-	OutArray.AddUninitialized( sizeof(double) * InArray.Num());
+	OutArray.AddUninitialized(sizeof(double) * InArray.Num());
 
 	TArray<uint8>::SizeType ByteIndex = DATA_OFFSET;
-	for(const auto& Number : InArray)
+	for (const auto& Number : InArray)
 	{
 		FMemory::Memcpy(OutArray.GetData() + ByteIndex, &Number, sizeof(double));
 		ByteIndex += sizeof(double);
@@ -48,7 +49,7 @@ static void ConvertArrayToBytes(const TArray<double>& InArray, TArray<uint8>& Ou
 
 static bool ConvertBytesToArray(const TArray<uint8>& InArray, TArray<FString>& OutArray)
 {
-	if(InArray.Num() < DATA_OFFSET)
+	if (InArray.Num() < DATA_OFFSET)
 	{
 		return false;
 	}
@@ -58,19 +59,19 @@ static bool ConvertBytesToArray(const TArray<uint8>& InArray, TArray<FString>& O
 
 	// Ensure that the data type field is correct, and that the number of bytes is evenly divisible by the number of
 	// bytes in a TCHAR to avoid reading outside of the bounds of InArray
-	if(DataType != ESessionSettingsAccelByteArrayFieldType::STRINGS || DataSize % sizeof(TCHAR) != 0)
+	if (DataType != ESessionSettingsAccelByteArrayFieldType::STRINGS || DataSize % sizeof(TCHAR) != 0)
 	{
 		return false;
 	}
 
 	FString String = TEXT("");
-	for(TArray<uint8>::SizeType i = DATA_OFFSET; i < DataSize + DATA_OFFSET; i += sizeof(TCHAR))
+	for (TArray<uint8>::SizeType i = DATA_OFFSET; i < DataSize + DATA_OFFSET; i += sizeof(TCHAR))
 	{
 		TCHAR Char;
 		FMemory::Memcpy(&Char, InArray.GetData() + i, sizeof(TCHAR));
 
 		// When serializing, we inserted a TCHAR with value 0 to separate the array items
-		if(Char == 0)
+		if (Char == 0)
 		{
 			OutArray.Add(String);
 			String = TEXT("");
@@ -86,7 +87,7 @@ static bool ConvertBytesToArray(const TArray<uint8>& InArray, TArray<FString>& O
 
 static bool ConvertBytesToArray(const TArray<uint8>& InArray, TArray<double>& OutArray)
 {
-	if(InArray.Num() < DATA_OFFSET)
+	if (InArray.Num() < DATA_OFFSET)
 	{
 		return false;
 	}
@@ -96,12 +97,12 @@ static bool ConvertBytesToArray(const TArray<uint8>& InArray, TArray<double>& Ou
 
 	// Ensure that the data type field is correct, and that the number of bytes is evenly divisible by the number of
 	// bytes in a double to avoid reading outside of the bounds of InArray
-	if(DataType != ESessionSettingsAccelByteArrayFieldType::DOUBLES || DataSize % sizeof(double) != 0)
+	if (DataType != ESessionSettingsAccelByteArrayFieldType::DOUBLES || DataSize % sizeof(double) != 0)
 	{
 		return false;
 	}
 
-	for(TArray<uint8>::SizeType i = DATA_OFFSET; i < DataSize + DATA_OFFSET; i += sizeof(double))
+	for (TArray<uint8>::SizeType i = DATA_OFFSET; i < DataSize + DATA_OFFSET; i += sizeof(double))
 	{
 		double Number;
 		FMemory::Memcpy(&Number, InArray.GetData() + i, sizeof(double));
@@ -140,7 +141,7 @@ void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, 
 	SearchSettings.Set(Key, ResultArray, InType);
 }
 
-void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, FName Key, const TArray<double>& Value,	EOnlineComparisonOp::Type InType)
+void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, FName Key, const TArray<double>& Value, EOnlineComparisonOp::Type InType)
 {
 	TArray<uint8> ResultArray;
 	ConvertArrayToBytes(Value, ResultArray);
@@ -154,7 +155,7 @@ void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, 
 	SearchSettings.Set(Key, ResultArray, InType, ID);
 }
 
-void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, FName Key, const TArray<double>& Value,	EOnlineComparisonOp::Type InType, int32 ID)
+void FOnlineSearchSettingsAccelByte::Set(FOnlineSearchSettings& SearchSettings, FName Key, const TArray<double>& Value, EOnlineComparisonOp::Type InType, int32 ID)
 {
 	TArray<uint8> ResultArray;
 	ConvertArrayToBytes(Value, ResultArray);
@@ -174,7 +175,7 @@ bool FOnlineSearchSettingsAccelByte::Get(FName Key, TArray<double>& Value) const
 bool FOnlineSearchSettingsAccelByte::Get(const FOnlineSearchSettings& SearchSettings, FName Key, TArray<FString>& Value)
 {
 	TArray<uint8> RawArray;
-	if(!SearchSettings.Get(Key, RawArray))
+	if (!SearchSettings.Get(Key, RawArray))
 	{
 		return false;
 	}
@@ -185,7 +186,7 @@ bool FOnlineSearchSettingsAccelByte::Get(const FOnlineSearchSettings& SearchSett
 bool FOnlineSearchSettingsAccelByte::Get(const FOnlineSearchSettings& SearchSettings, FName Key, TArray<double>& Value)
 {
 	TArray<uint8> RawArray;
-	if(!SearchSettings.Get(Key, RawArray))
+	if (!SearchSettings.Get(Key, RawArray))
 	{
 		return false;
 	}
@@ -207,10 +208,39 @@ bool FOnlineSearchSettingsAccelByte::Get(const FVariantData& Data, TArray<double
 	return ConvertBytesToArray(RawArray, Value);
 }
 
+template<typename T>
+bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, T& OutValue)
+{
+	double DoubleValue = -1.0;
+	if (!SearchSettings.Get(Key, DoubleValue))
+	{
+		return false;
+	}
+
+	if (DoubleValue <= static_cast<double>(TNumericLimits<T>::Min()) || DoubleValue >= static_cast<double>(TNumericLimits<T>::Max()))
+	{
+		return false;
+	}
+
+	OutValue = static_cast<T>(FMath::RoundHalfFromZero(DoubleValue));
+	return true;
+}
+
+// Begin specialized GetInt definitions
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, int8& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, int16& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, int32& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, int64& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, uint8& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, uint16& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, uint32& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSearchSettingsAccelByte::GetInt(const FOnlineSearchSettings& SearchSettings, FName Key, uint64& OutValue);
+// End specialized GetInt definitions
+
 ESessionSettingsAccelByteArrayFieldType FOnlineSearchSettingsAccelByte::GetArrayFieldType(const FOnlineSearchSettings& SearchSettings, FName Key)
 {
 	TArray<uint8> RawArray;
-	if(!SearchSettings.Get(Key, RawArray) || RawArray.Num() < DATA_OFFSET)
+	if (!SearchSettings.Get(Key, RawArray) || RawArray.Num() < DATA_OFFSET)
 	{
 		return ESessionSettingsAccelByteArrayFieldType::INVALID;
 	}
@@ -222,7 +252,7 @@ ESessionSettingsAccelByteArrayFieldType FOnlineSearchSettingsAccelByte::GetArray
 {
 	TArray<uint8> RawArray;
 	Data.GetValue(RawArray);
-	if(RawArray.Num() < DATA_OFFSET)
+	if (RawArray.Num() < DATA_OFFSET)
 	{
 		return ESessionSettingsAccelByteArrayFieldType::INVALID;
 	}
@@ -315,7 +345,7 @@ bool FOnlineSessionSettingsAccelByte::Get(FName Key, TArray<double>& OutValue) c
 bool FOnlineSessionSettingsAccelByte::Get(const FOnlineSessionSettings& SessionSettings, FName Key, TArray<FString>& OutValue)
 {
 	TArray<uint8> RawArray;
-	if(!SessionSettings.Get(Key, RawArray))
+	if (!SessionSettings.Get(Key, RawArray))
 	{
 		return false;
 	}
@@ -326,7 +356,7 @@ bool FOnlineSessionSettingsAccelByte::Get(const FOnlineSessionSettings& SessionS
 bool FOnlineSessionSettingsAccelByte::Get(const FOnlineSessionSettings& SessionSettings, FName Key, TArray<double>& OutValue)
 {
 	TArray<uint8> RawArray;
-	if(!SessionSettings.Get(Key, RawArray))
+	if (!SessionSettings.Get(Key, RawArray))
 	{
 		return false;
 	}
@@ -334,10 +364,39 @@ bool FOnlineSessionSettingsAccelByte::Get(const FOnlineSessionSettings& SessionS
 	return ConvertBytesToArray(RawArray, OutValue);
 }
 
+template<typename T>
+bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, T& OutValue)
+{
+	double DoubleValue = -1.0;
+	if (!SessionSettings.Get(Key, DoubleValue))
+	{
+		return false;
+	}
+
+	if (DoubleValue <= static_cast<double>(TNumericLimits<T>::Min()) || DoubleValue >= static_cast<double>(TNumericLimits<T>::Max()))
+	{
+		return false;
+	}
+
+	OutValue = static_cast<T>(FMath::RoundHalfFromZero(DoubleValue));
+	return true;
+}
+
+// Begin specialized GetInt definitions
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, int8& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, int16& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, int32& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, int64& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, uint8& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, uint16& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, uint32& OutValue);
+template ONLINESUBSYSTEMACCELBYTE_API bool FOnlineSessionSettingsAccelByte::GetInt(const FOnlineSessionSettings& SessionSettings, FName Key, uint64& OutValue);
+// End specialized GetInt definitions
+
 ESessionSettingsAccelByteArrayFieldType FOnlineSessionSettingsAccelByte::GetArrayFieldType(const FOnlineSessionSettings& SessionSettings, FName Key)
 {
 	TArray<uint8> RawArray;
-	if(!SessionSettings.Get(Key, RawArray) || RawArray.Num() < DATA_OFFSET)
+	if (!SessionSettings.Get(Key, RawArray) || RawArray.Num() < DATA_OFFSET)
 	{
 		return ESessionSettingsAccelByteArrayFieldType::INVALID;
 	}